Disable SSL certificate hostname check
We need to connect to exact IP to perform HTTPS tests. There is no standard way to redirect
one hostname to another IP address so we connect to the IP address and add proper "Host"
header later. This causes SSL hostname validation error. There's a little point in it
for our needs at all so it's safe to disable it completely.